Grammar fixes for Plugins-Guide.md
Only a few of these grammar issues made it hard to read, but I went ahead and fixed all that I could find in a quick scan.

This documentation is otherwise a fun and easy read!

BTW, here in the U.S. women have become pretty sensitive to the use of masculine terms when gender-neutral terms exist. I suspect saying "children" and "siblings" would better suit modern sensitivities than "sons" and "brothers." But that's not a grammar issue, so I didn't change it. Just a heads-up.