feat: make version field deterministic and reliable (#3427)
* feat: make version field deterministic and reliable

This change makes the "version" a reliable field, instead of
one that might be "undefined". Some issues with bundlers such
as "a wild package.json appears" are resolved as well.

Maintainers would have to ensure that versions in fastify.js
and package.json are in sync. That would be relatively trivial,
compared to the benefits, when it comes to such an essential
package.

To avoid errors and reduce maintenance burden, a test has been
added to ensure that the versions are in sync. It is configured
to run before publish.

Refs: #3110, #3113, #3155, #3185, #3191
Signed-off-by: Jesse Chan <jc@linux.com>

* test: bundler: drop no longer relevant version fallback test

Signed-off-by: Jesse Chan <jc@linux.com>

* build: add a script to sync versions

Signed-off-by: Jesse Chan <jc@linux.com>

* test: bundler: restore version test but invert assertion

Signed-off-by: Jesse Chan <jc@linux.com>